rename: 패키지 구조 도메인별로 나누는 방식으로 변경
